divide off undo_redo
